Complete 06-flex-layout
Updates style.css to add flexbox to header, content,
and footer.
Note: My solution was slightly different adding flexbox
to all three elements and not adding to body element.
Because of this I had to add more padding and margins.
Solution for links was slightly different as well using
something I found online.